crop-prediction/backend/requirement.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# AI & Image Processing
tensorflow-cpu>=2.15.0
numpy>=1.23.5
pillow>=10.2.0

# Database & Utilities
psutil==5.9.8
requests==2.31.0